update the source from root pom and compile project
