+++ .idea/encodings.xml
@@ -3,5 +3,7 @@
   <component name="Encoding">
     <file url="file://$PROJECT_DIR$/src/main/java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/src/main/resources" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/toolkits/codebank/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/toolkits/discourse/src/main/java" charset="UTF-8" />
   </component>
 </project>
