--- src/main/java/com/ossez/codebank/algorithm/models/ListNode.java
@@ -1,16 +1,0 @@
-package com.ossez.codebank.algorithm.models;
-
-/**
- * 
- * @author YuCheng
- *
- */
-public class ListNode {
-	public  int val;
-	public  ListNode next;
-
-	public ListNode(int x) {
-		val = x;
-		next = null;
-	}
-}
