--- src/main/java/com/ossez/codebank/algorithm/models/TreeNode.java
@@ -1,16 +1,0 @@
-package com.ossez.codebank.algorithm.models;
-
-/**
- * 
- * @author YuCheng
- *
- */
-public class TreeNode {
-	public int val;
-	public TreeNode left, right;
-
-	public TreeNode(int val) {
-		this.val = val;
-		this.left = this.right = null;
-	}
-}
