--- src/test/java/com/ossez/codebank/algorithm/tests/lintcode/LintCode1480DotProductTest.java
@@ -1,35 +1,0 @@
-package com.ossez.codebank.algorithm.tests.lintcode;
-
-import org.junit.Test;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-/**
- * <p>
- * 1480
- * <ul>
- * <li>@see
- * <a href= "https://www.cwiki.us/display/ITCLASSIFICATION/Dot+Product">https://www.cwiki.us/display/ITCLASSIFICATION/Dot+Product</a>
- * <li>@see<a href= "https://www.lintcode.com/problem/dot-product/">https://www.lintcode.com/problem/dot-product/</a>
- * </ul>
- * </p>
- * 
- * @author YuCheng
- *
- */
-public class LintCode1480DotProductTest {
-
-	private final static Logger logger = LoggerFactory.getLogger(LintCode1480DotProductTest.class);
-
-	/**
-	 * 
-	 */
-	@Test
-	public void testMain() {
-		logger.debug("BEGIN");
-		int t = 87;
-		int[] dur = { 20, 25, 19, 37 };
-		// Write your code here
-	}
-
-}
