--- toolkits/.idea/compiler.xml
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="CompilerConfiguration">
-    <annotationProcessing>
-      <profile name="Maven default annotation processors profile" enabled="true">
-        <sourceOutputDir name="target/generated-sources/annotations" />
-        <sourceTestOutputDir name="target/generated-test-sources/test-annotations" />
-        <outputRelativeToContentRoot value="true" />
-        <module name="codebank" />
-        <module name="discourse" />
-      </profile>
-    </annotationProcessing>
-    <bytecodeTargetLevel>
-      <module name="codebank" target="11" />
-      <module name="discourse" target="11" />
-      <module name="toolkits" target="11" />
-    </bytecodeTargetLevel>
-  </component>
-</project>
