+++ toolkits/codebank/src/main/java/com/ossez/toolkits/codebank/common/interview/ManNextFibonacciNumber.java
@@ -1,4 +1,4 @@
-package com.ossez.codebank.interview;
+package com.ossez.toolkits.codebank.common.interview;
 
 import java.io.BufferedReader;
 import java.io.InputStreamReader;
