+++ toolkits/codebank/src/main/java/com/ossez/toolkits/codebank/common/usecase/VarargsCase.java
