+++ toolkits/codebank/src/test/java/com/ossez/toolkits/codebank/tests/LintcodeTest.java
@@ -1,17 +1,11 @@
-package com.ossez.lang.tutorial.tests;
+package com.ossez.toolkits.codebank.tests;
 
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Collections;
-import java.util.HashMap;
 import java.util.HashSet;
-import java.util.LinkedList;
 import java.util.List;
-import java.util.TreeSet;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
-
-import javax.management.ListenerNotFoundException;
 
 import org.junit.Test;
 import org.slf4j.Logger;
