+++ toolkits/codebank/src/test/java/com/ossez/toolkits/codebank/tests/interview/LambdaEvensTest.java
@@ -1,4 +1,4 @@
-package com.ossez.codebank.interview.tests.others;
+package com.ossez.toolkits.codebank.tests.interview;
 
 import org.junit.Test;
 import org.slf4j.Logger;
